feat: refactor interceptor imports for consistency across controllers
