+++ src/controllers/health.controller.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from "express";
 import { IServiceResponse } from "../interface/common.interface";
 import APIConfig from "../utils/config";
-import ResponseInterceptor from "../interceptors/response.interceptor";
+import { ResponseInterceptor } from "../interceptors";
 
 const serviceName = APIConfig.config.serviceName;
