+++ src/controllers/login.controller.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from "express";
 import loginService from "../services/login.service";
 import { clearCookies } from "../utils/common.util";
-import ResponseInterceptor from "../interceptors/response.interceptor";
+import { ResponseInterceptor } from "../interceptors";
 
 class LoginController {
   public signUp(req: Request, res: Response, next: NextFunction) {
