+++ src/express-app.ts
@@ -3,8 +3,7 @@ import express, { Express } from "express";
 import MongoConnection from "./db/mongo-connection";
 import Routes from "./routers/router";
 import APIConfig from "./utils/config";
-import ErrorInterceptor from "./interceptors/error.interceptor";
-import RequestInterceptor from "./interceptors/request.interceptor";
+import { RequestInterceptor, ErrorInterceptor } from "./interceptors";
 import SecurityMiddleware from "./middleware/security.middleware";
 import { watchUserCollection } from "./services/change-stream.service";
 import swaggerUi from "swagger-ui-express";
